move service provider comparator to own file
